split in sub components
