+++ addons/compat_xeh_fix/compat_breach/script_component.hpp
@@ -1,0 +1,4 @@
+#define SUBCOMPONENT breach
+#define SUBCOMPONENT_BEAUTIFIED Breach
+
+#include "..\script_component.hpp"
