+++ addons/compat_xeh_fix/compat_bwmod/CfgVehicles.hpp
@@ -1,0 +1,5 @@
+class CfgVehicles {
+    class Mortar_01_base_F;
+    class BWA3_MRS120_base: Mortar_01_base_F { XEH_ENABLED; }; // "Mk6 Mortar" @BWMod
+};
+
