+++ addons/compat_xeh_fix/compat_bwmod/readme.md
@@ -1,0 +1,5 @@
+# XEH Fix - BWA3
+
+## Maintainer
+
+- Andx
