+++ addons/compat_xeh_fix/compat_bwmod/script_component.hpp
@@ -1,0 +1,4 @@
+#define SUBCOMPONENT bwa3
+#define SUBCOMPONENT_BEAUTIFIED BWA3
+
+#include "..\script_component.hpp"
