+++ addons/compat_xeh_fix/compat_fir_aws/config.cpp
@@ -1,15 +1,17 @@
 #include "script_component.hpp"
 
 class CfgPatches {
-    class ADDON {
+    class SUBADDON {
         name = COMPONENT_NAME;
         units[] = {};
         weapons[] = {};
         requiredVersion = REQUIRED_VERSION;
-        requiredAddons[] = {"ttt_main"};
+        requiredAddons[] = {"FIR_AirWeaponSystem_US","A3_Static_F_Sams_Radar_System_01","FIR_DDGLibertyUpgrade_F"};
         author = ECSTRING(main,TacticalTrainingTeam);
         authors[] = {""};
         url = ECSTRING(main,URL);
         VERSION_CONFIG;
     };
 };
+
+#include "CfgVehicles.hpp"
