+++ addons/compat_xeh_fix/compat_fir_aws/readme.md
@@ -1,0 +1,5 @@
+# XEH Fix - FIR AWS
+
+## Maintainer
+
+- Andx
