+++ addons/compat_xeh_fix/compat_fir_aws/script_component.hpp
@@ -1,0 +1,4 @@
+#define SUBCOMPONENT firaws
+#define SUBCOMPONENT_BEAUTIFIED FIR AWS
+
+#include "..\script_component.hpp"
