+++ addons/compat_xeh_fix/compat_pbw/CfgVehicles.hpp
@@ -1,0 +1,6 @@
+class CfgVehicles {
+    class ReammoBox_F;
+    class PBW_Uniform1_Chest: ReammoBox_F { XEH_ENABLED; }; // "[PBW] Uniform Chest" @Bundeswehr_Kleiderkammer_PBW
+    class PBW_Headgear_Chest: PBW_Uniform1_Chest { XEH_ENABLED; }; // "[PBW] Headgear Chest" @Bundeswehr_Kleiderkammer_PBW
+};
+
