+++ addons/compat_xeh_fix/compat_pbw/readme.md
@@ -1,0 +1,5 @@
+# XEH Fix - PBW
+
+## Maintainer
+
+- Andx
