+++ addons/compat_xeh_fix/compat_pbw/script_component.hpp
@@ -1,0 +1,4 @@
+#define SUBCOMPONENT pbw
+#define SUBCOMPONENT_BEAUTIFIED PBW
+
+#include "..\script_component.hpp"
