+++ addons/compat_xeh_fix/compat_rnt/CfgVehicles.hpp
@@ -1,0 +1,8 @@
+class CfgVehicles {
+    class GMG_01_base_F;
+    class rnt_gmw_static: GMG_01_base_F { XEH_ENABLED; }; // "GMW (Tripod)" @Redd_n_Tank_Vehicles
+
+    class HMG_01_high_base_F;
+    class rnt_mg3_static: HMG_01_high_base_F { XEH_ENABLED; }; // "MG3 (Tripod)" @Redd_n_Tank_Vehicles
+};
+
