+++ addons/compat_xeh_fix/compat_rnt/config.cpp
@@ -6,10 +6,12 @@ class CfgPatches {
         units[] = {};
         weapons[] = {};
         requiredVersion = REQUIRED_VERSION;
-        requiredAddons[] = {"ttt_main"};
+        requiredAddons[] = {"Redd_Tank_M120_Tampella","m120_comp_ace","rnt_ace_compatibility","rnt_gmw_static","rnt_mg3_static"};
         author = ECSTRING(main,TacticalTrainingTeam);
         authors[] = {""};
         url = ECSTRING(main,URL);
         VERSION_CONFIG;
     };
 };
+
+#include "CfgVehicles.hpp"
