+++ addons/compat_xeh_fix/compat_rnt/readme.md
@@ -1,0 +1,5 @@
+# XEH Fix Redd n Tank
+
+## Maintainer
+
+- Andx
