+++ addons/compat_xeh_fix/compat_rnt/script_component.hpp
@@ -1,0 +1,3 @@
+#define SUBCOMPONENT rnt
+#define SUBCOMPONENT_BEAUTIFIED Redd n Tank
+#include "..\script_component.hpp"
