+++ addons/compat_xeh_fix/compat_vanilla/CfgVehicles.hpp
@@ -1,0 +1,8 @@
+class CfgVehicles {
+    class Radar_System_01_base_F;
+    class B_Radar_System_01_F: Radar_System_01_base_F { XEH_ENABLED; }; // "AN/MPQ-105 Radar" jets
+
+    class Radar_System_02_base_F;
+    class O_Radar_System_02_F: Radar_System_02_base_F { XEH_ENABLED; }; // "R-750 Cronus Radar" jets
+};
+
