+++ addons/compat_xeh_fix/compat_vanilla/config.cpp
@@ -6,10 +6,12 @@ class CfgPatches {
         units[] = {};
         weapons[] = {};
         requiredVersion = REQUIRED_VERSION;
-        requiredAddons[] = {"ttt_main"};
+        requiredAddons[] = {"A3_Static_F_Sams_Radar_System_01","A3_Static_F_Sams_Radar_System_02"};
         author = ECSTRING(main,TacticalTrainingTeam);
         authors[] = {""};
         url = ECSTRING(main,URL);
         VERSION_CONFIG;
     };
 };
+
+#include "CfgVehicles.hpp"
