+++ addons/compat_xeh_fix/compat_vanilla/readme.md
@@ -1,0 +1,5 @@
+# XEH Fix - Vanilla
+
+## Maintainer
+
+- Andx
