+++ addons/compat_xeh_fix/compat_vanilla/script_component.hpp
@@ -1,0 +1,3 @@
+#define SUBCOMPONENT vanilla
+#define SUBCOMPONENT_BEAUTIFIED Vanilla
+#include "..\script_component.hpp"
