+++ addons/compat_xeh_fix/readme.md
@@ -1,4 +1,4 @@
-# Fix Extended Event Handlers
+# XEH Fix
 
 ## Referenzen
